+++ packages/source-filesystem/index.js
@@ -1,5 +1,6 @@
 const path = require('path')
 const fs = require('fs-extra')
+const slash = require('slash')
 const { mapValues } = require('lodash')
 
 const isDev = process.env.NODE_ENV === 'development'
@@ -86,7 +87,6 @@ class FilesystemSource {
   }
 
   watchFiles () {
-    const slash = require('slash')
     const chokidar = require('chokidar')
 
     const watcher = chokidar.watch(this.options.path, {
@@ -158,11 +158,12 @@ class FilesystemSource {
   }
 
   createPath ({ dir, name }) {
-    const { route, pathPrefix = '' } = this.options
+    const { route, pathPrefix = '/' } = this.options
 
     if (route) return
 
-    const segments = (pathPrefix + dir)
+    const joinedPath = path.join(pathPrefix, dir)
+    const segments = slash(joinedPath)
       .split('/')
       .filter(v => v)
       .map(s => this.store.slugify(s))