+++ packages/source-drupal/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.1](https://github.com/gridsome/gridsome/tree/master/packages/source-drupal/compare/@gridsome/source-drupal@0.2.0...@gridsome/source-drupal@0.2.1) (2019-05-20)
+
+**Note:** Version bump only for package @gridsome/source-drupal
+
+
+
+
+
 # [0.2.0](https://github.com/gridsome/gridsome/tree/master/packages/source-drupal/compare/@gridsome/source-drupal@0.1.1...@gridsome/source-drupal@0.2.0) (2019-05-10)